Add copy of Apache License
This adds a copy of the Apache license to this repository;

    curl https://www.apache.org/licenses/LICENSE-2.0.txt > LICENSE

Note that the copyright at the end of the license should be left
blank, as it's only a template/example that can be used for adding
copyright headers to source files. The license itself should be
included as-is (without modifications).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>